backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.23.3
pydantic==2.11.7
pydantic-settings==2.9.1
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
openai==1.3.7
anthropic==0.7.7
solana==0.32.0
anyio==4.9.0
typing-extensions==4.14.0
typing-inspection==0.4.1
